+++ transport/http/inbound.go
@@ -24,6 +24,7 @@ import (
 	"context"
 	"crypto/tls"
 	"errors"
+	"fmt"
 	"net"
 	"net/http"
 	"strings"
@@ -38,15 +39,19 @@ import (
 	"go.uber.org/yarpc/transport/internal/tls/muxlistener"
 	"go.uber.org/yarpc/yarpcerrors"
 	"go.uber.org/zap"
+	"golang.org/x/net/http2"
+	"golang.org/x/net/http2/h2c"
 )
 
-// We want a value that's around 5 seconds, but slightly higher than how
-// long a successful HTTP shutdown can take.
-// There's a specific path in the HTTP shutdown path that can take 5 seconds:
-// https://golang.org/src/net/http/server.go?s=83923:83977#L2710
-// This avoids timeouts in shutdown caused by new idle connections, without
-// making the timeout too large.
-const defaultShutdownTimeout = 6 * time.Second
+const (
+	// We want a value that's around 5 seconds, but slightly higher than how
+	// long a successful HTTP shutdown can take.
+	// There's a specific path in the HTTP shutdown path that can take 5 seconds:
+	// https://golang.org/src/net/http/server.go?s=83923:83977#L2710
+	// This avoids timeouts in shutdown caused by new idle connections, without
+	// making the timeout too large.
+	defaultShutdownTimeout = 6 * time.Second
+)
 
 // InboundOption customizes the behavior of an HTTP Inbound constructed with
 // NewInbound.
@@ -124,6 +129,13 @@ func InboundTLSMode(mode yarpctls.Mode) InboundOption {
 	}
 }
 
+// DisableHTTP2 returns an InboundOption that disables HTTP/2 support.
+func DisableHTTP2(flag bool) InboundOption {
+	return func(i *Inbound) {
+		i.disableHTTP2 = flag
+	}
+}
+
 // NewInbound builds a new HTTP inbound that listens on the given address and
 // sharing this transport.
 func (t *Transport) NewInbound(addr string, opts ...InboundOption) *Inbound {
@@ -136,6 +148,7 @@ func (t *Transport) NewInbound(addr string, opts ...InboundOption) *Inbound {
 		transport:         t,
 		grabHeaders:       make(map[string]struct{}),
 		bothResponseError: true,
+		disableHTTP2:      false,
 	}
 	for _, opt := range opts {
 		opt(i)
@@ -165,6 +178,8 @@ type Inbound struct {
 
 	tlsConfig *tls.Config
 	tlsMode   yarpctls.Mode
+
+	disableHTTP2 bool
 }
 
 // Tracer configures a tracer on this inbound.
@@ -221,10 +236,19 @@ func (i *Inbound) start() error {
 		httpHandler = i.mux
 	}
 
-	i.server = intnet.NewHTTPServer(&http.Server{
+	server := &http.Server{
 		Addr:    i.addr,
 		Handler: httpHandler,
-	})
+	}
+	if !i.disableHTTP2 {
+		h2s := &http2.Server{}
+		server.Handler = h2c.NewHandler(server.Handler, h2s)
+		err := http2.ConfigureServer(server, h2s)
+		if err != nil {
+			return fmt.Errorf("failed to configure HTTP/2 server: %w", err)
+		}
+	}
+	i.server = intnet.NewHTTPServer(server)
 
 	addr := i.addr
 	if addr == "" {